RP-12 패키지 구조 변경 완료
