+++ src/main/java/dev/teamkjk/reservation_system/domain/performance/ConcertEntity.java
@@ -1,4 +1,4 @@
-package dev.teamkjk.reservation_system.persistence.entity.performance;
+package dev.teamkjk.reservation_system.domain.performance;
 
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
